Refactor workflow canvas helpers into modules
